feat: improve space API typings  (#684)
BREAKING CHANGE: old typings were too loose.
